Bug mensajes repetidos al actualizad (Solucionado)

+++ django-chat/chat/views.py
@@ -9,6 +9,17 @@ def lobby(request):
     all_Post = Post.objects.all()
     return render(request, 'chat/lobby.html',{'mensajes':all_Post})
 def aEliminar(request, uuid=None):
-    postAEliminar = Post.objects.get(uuid=uuid) 
-    postAEliminar.delete()
+    if(uuid!='def'):
+        try:
+            postAEliminar = Post.objects.get(uuid=uuid) 
+            postAEliminar.delete()
+        except:
+            pass
+    return redirect('/')
+def aAgregar(request):
+    if(request.method == 'POST'):
+        message = request.POST['message']
+        mensaje = Post.objects.create(text= message, author = str(request.user))
+        mensaje.publish()
+  
     return redirect('/')
